# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wapul-be",
    "version": "a1e44bde9c77"
  },
  "paths": {
...
    "/v1/records/{record_id}/groups": {
      "put": {
        "tags": [
          "Records"
        ],
        "summary": "Update Shares",
        "description": "기록을 공유할 그룹을 통째로 정한다. 내가 속한 그룹에만 공유할 수 있다.",
        "operationId": "update_shares_v1_records__record_id__groups_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SharesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "title": "Response Update Shares V1 Records  Record Id  Groups Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/groups": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "List My Groups",
        "operationId": "list_my_groups_v1_groups_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GroupOut"
                  },
                  "type": "array",
                  "title": "Response List My Groups V1 Groups Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Create Group",
        "operationId": "create_group_v1_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/groups/join": {
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Join Group",
        "operationId": "join_group_v1_groups_join_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupJoin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/groups/{group_id}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Get Group",
        "operationId": "get_group_v1_groups__group_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GroupCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "GroupCreate"
      },
      "GroupDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "invite_code": {
            "type": "string",
            "title": "Invite Code"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/MemberOut"
            },
            "type": "array",
            "title": "Members"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/SharedRecordOut"
            },
            "type": "array",
            "title": "Records"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "invite_code",
          "role",
          "member_count",
          "members",
          "records"
        ],
        "title": "GroupDetail"
      },
      "GroupJoin": {
        "properties": {
          "invite_code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Invite Code"
          }
        },
        "type": "object",
        "required": [
          "invite_code"
        ],
        "title": "GroupJoin"
      },
      "GroupOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "invite_code": {
            "type": "string",
            "title": "Invite Code"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "invite_code",
          "role",
          "member_count"
        ],
        "title": "GroupOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Language": {
        "type": "string",
        "enum": [
          "cpp",
          "python",
          "java"
        ],
        "title": "Language"
      },
...
      "MemberOut": {
        "properties": {
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "display_name",
          "role"
        ],
        "title": "MemberOut"
      },
...
      "SharedRecordOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "key_idea": {
            "type": "string",
            "title": "Key Idea"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "shared_at": {
            "type": "string",
            "format": "date-time",
            "title": "Shared At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "problem",
          "key_idea",
          "language",
          "owner_name",
          "created_at",
          "updated_at",
          "shared_at"
        ],
        "title": "SharedRecordOut"
      },
      "SharesUpdate": {
        "properties": {
          "group_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Group Ids"
          }
        },
        "type": "object",
        "required": [
          "group_ids"
        ],
        "title": "SharesUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}